+++ CHANGES.rst
@@ -6,6 +6,7 @@ Version 1.0.1
 
 * Added ``marker_zoom`` setting to clusters
 * Improve visual display of marker groups.
+* Improve jQuery plugin interface, allow method calling with ``.mapPlugin("methodname", args..)``.
 
 Version 1.0 (2015-06-09)
 ------------------------